reorganize --first pass-- app is working, tests failing
