--- mathesar/database/types/constants.py
@@ -1,4 +1,0 @@
-from mathesar.settings import mathesar_settings
-
-APP_PREFIX = mathesar_settings["APP_PREFIX"]
-TYPE_SCHEMA = f"{APP_PREFIX}types"
