+++ mathesar/forms/forms.py
@@ -1,9 +1,9 @@
 from django import forms
 from django.core.exceptions import ValidationError
 
-from mathesar.database.schemas import get_all_schemas
 from mathesar.database.base import create_mathesar_engine
 from mathesar.forms.widgets import DataListInput
+from mathesar_db import schemas
 
 
 engine = create_mathesar_engine()
@@ -19,7 +19,7 @@ class UploadFileForm(forms.Form):
 
     application_name = forms.CharField(
         min_length=1, label="Application Name",
-        widget=DataListInput(get_all_schemas(engine))
+        widget=DataListInput(schemas.get_all_schemas(engine))
     )
 
     file = forms.FileField(validators=[validate_csv], label="CSV File")
