+++ mathesar/imports/csv.py
@@ -2,8 +2,8 @@ import csv
 from io import TextIOWrapper
 
 from mathesar.database.base import create_mathesar_engine
-from mathesar.database.tables import create_table, insert_rows_into_table
 from mathesar.models import Collection
+from mathesar_db import tables
 
 engine = create_mathesar_engine()
 
@@ -15,8 +15,8 @@ def get_csv_reader(csv_file):
 
 
 def create_table_from_csv(name, schema, csv_reader, engine):
-    table = create_table(name, schema, csv_reader.fieldnames, engine)
-    insert_rows_into_table(table, [row for row in csv_reader], engine)
+    table = tables.create_table(name, schema, csv_reader.fieldnames, engine)
+    tables.insert_rows_into_table(table, [row for row in csv_reader], engine)
     return table
