+++ mathesar_db/constants.py
@@ -1,0 +1,2 @@
+MATHESAR_PREFIX = "mathesar_"
+ID = f"{MATHESAR_PREFIX}id"
