+++ mathesar_db/types/__init__.py
@@ -1,4 +1,4 @@
-from mathesar.database.types import email
+from mathesar_db.types import email
 
 CUSTOM_TYPE_DICT = {
     email.QUALIFIED_EMAIL: email.Email,
