+++ mathesar_db/types/base.py
@@ -1,0 +1,10 @@
+from sqlalchemy import create_engine
+from mathesar_db import constants
+
+SCHEMA = f"{constants.MATHESAR_PREFIX}types"
+# Since we want to have our identifiers quoted appropriately for use in
+# PostgreSQL, we want to use the postgres dialect preparer to set this up.
+preparer = create_engine("postgresql://").dialect.identifier_preparer
+
+def get_qualified_name(name):
+    return ".".join([preparer.quote_schema(SCHEMA), name])
